sdk: renamend & moved StResponse to api package and renamed to ApiResponse
