+++ sdk/src/main/java/com/sygic/travel/sdk/common/api/model/ApiResponse.kt
@@ -1,4 +1,4 @@
-package com.sygic.travel.sdk.common.responseWrappers
+package com.sygic.travel.sdk.common.api.model
 
 import com.google.gson.annotations.SerializedName
 
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
  * Base of an API response. It's extended by specific respond classes containing the specific data.
  * It's abstract, because of the abstract [.getData] method.
  */
-abstract class StResponse {
+abstract class ApiResponse {
 	@SerializedName("status_code")
 	val statusCode: Int = 0
