+++ sdk/src/main/java/com/sygic/travel/sdk/places/responseWrappers/PlacesApiResponse.kt
@@ -1,15 +1,15 @@
 package com.sygic.travel.sdk.places.responseWrappers
 
 import com.google.gson.annotations.SerializedName
-import com.sygic.travel.sdk.common.responseWrappers.StResponse
+import com.sygic.travel.sdk.common.api.model.ApiResponse
 import com.sygic.travel.sdk.places.model.Place
 import com.sygic.travel.sdk.places.model.api.place.ApiPlaceDetail
 
 /**
- * Response that contains a list of places data. Suitable for showing places in a list or
+ * ApiResponse that contains a list of places data. Suitable for showing places in a list or
  * spread on a map.
  */
-class PlacesResponse : StResponse() {
+class PlacesApiResponse : ApiResponse() {
 	private var data: Data? = null
 
 	fun getPlaces(): List<Place>? {
