+++ sdk/src/main/java/com/sygic/travel/sdk/tours/responseWrapper/TourApiResponse.kt
@@ -1,14 +1,14 @@
 package com.sygic.travel.sdk.tours.responseWrapper
 
 import com.google.gson.annotations.SerializedName
-import com.sygic.travel.sdk.common.responseWrappers.StResponse
+import com.sygic.travel.sdk.common.api.model.ApiResponse
 import com.sygic.travel.sdk.tours.model.Tour
 import com.sygic.travel.sdk.tours.model.api.ApiTour
 
 /**
- * Response that contains a list of Tour classes.
+ * ApiResponse that contains a list of Tour classes.
  */
-class TourResponse : StResponse() {
+class TourApiResponse : ApiResponse() {
 	private var data: Data? = null
 
 	fun getTours(): List<Tour>? {
